Break out files in mod.
