+++ rust_dev_preview/cross_service/rest_ses/src/healthz.rs
@@ -1,0 +1,5 @@
+use actix_web::{HttpResponse, Responder};
+
+pub async fn healthz() -> impl Responder {
+    HttpResponse::Ok().body("OK\n")
+}
