+++ rust_dev_preview/cross_service/rest_ses/src/lib.rs
@@ -1,48 +1,22 @@
-use std::net::TcpListener;
+pub mod healthz;
+pub mod work_item;
 
-use serde::Deserialize;
+use std::net::TcpListener;
 
 use actix_web::{
     dev::Server,
-    web::{get, post, Json},
-    App, HttpResponse, HttpServer, Responder,
+    web::{get, scope},
+    App, HttpServer,
 };
-use uuid::Uuid;
 
-#[derive(Debug, Default, Deserialize)]
-struct WorkItem {
-    #[serde(default = "uuid::Uuid::new_v4")]
-    pub idwork: Uuid,
-    #[serde(default)]
-    pub username: String,
-    // date - A date value that specifies the date the item was created.
-    #[serde(default)]
-    pub description: String,
-    #[serde(default)]
-    pub guide: String,
-    #[serde(default)]
-    pub status: String,
-    #[serde(default)]
-    pub archive: u8,
-}
-
-async fn healthz() -> impl Responder {
-    HttpResponse::Ok().body("OK\n")
-}
-
-async fn create(item: Json<WorkItem>) -> impl Responder {
-    if item.username.is_empty() {
-        HttpResponse::BadRequest().body("Missing username")
-    } else {
-        HttpResponse::Ok().body(format!("adding item for {user}", user = item.username))
-    }
-}
+use healthz::healthz;
+use work_item::work_item_resource;
 
 pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
     let server = HttpServer::new(|| {
         App::new()
             .route("/healthz", get().to(healthz))
-            .route("/api/items", post().to(create))
+            .service(scope("/api").service(work_item_resource()))
     })
     .listen(listener)?
     .run();
